:recycle: refactor(Coupon): 도메인 모델 단위 테스트 코드 가독성 보완
- 성공/실패 여부 표기
- 테스트 대상 메소드별 묶음 처리
